【开发】增加quicklist 测试案例

+++ quicklist/demo_quicklist_2_endianconv.h
@@ -38,7 +38,7 @@ void memrev16(void *p);
 void memrev32(void *p);
 void memrev64(void *p);
 
-uint16_t intrev16(uint64_t v);
+uint16_t intrev16(uint16_t v);
 uint32_t intrev32(uint32_t v);
 uint64_t intrev64(uint64_t v);
